feat: rewrite npm-registry-agent to typescript
