--- packages/npm-registry-agent/index.d.ts
@@ -1,1 +1,0 @@
-export default Function
