+++ packages/npm-registry-agent/package.json
@@ -2,23 +2,40 @@
   "name": "@pnpm/npm-registry-agent",
   "version": "0.0.0",
   "description": "HTTP agent for fetching from npm registry",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
   "files": [
-    "index.js",
-    "index.d.ts"
+    "lib"
   ],
   "author": "Zoltan Kochan <z@kochan.io> (https://www.kochan.io/)",
   "license": "MIT",
   "scripts": {
-    "test": "node test"
+    "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
+    "tsc": "rimraf lib && tsc",
+    "test": "npm run tsc && npm run lint && ts-node test --type-check",
+    "prepublishOnly": "npm run tsc",
+    "fix": "tslint -c tslint.json src/**/*.ts test/**/*.ts --fix"
   },
   "devDependencies": {
+    "@pnpm/tslint-config": "0.0.0",
+    "@types/proxyquire": "1.3.28",
+    "@types/tape": "4.2.33",
     "proxyquire": "2.1.0",
-    "tape": "4.9.2"
+    "rimraf": "2.6.3",
+    "tape": "4.9.2",
+    "ts-node": "8.0.2",
+    "tslint": "5.12.1",
+    "typescript": "3.2.4"
   },
   "dependencies": {
+    "@types/http-proxy-agent": "2.0.1",
+    "@types/lru-cache": "4.1.1",
+    "@types/node": "10.12.18",
     "agentkeepalive": "4.0.0",
+    "http-proxy-agent": "2.1.0",
     "https-proxy-agent": "2.2.1",
-    "lru-cache": "5.1.1"
+    "lru-cache": "5.1.1",
+    "socks-proxy-agent": "4.0.1"
   },
   "env": {
     "node": ">=8"
