+++ packages/npm-registry-agent/test/index.ts
@@ -1,17 +1,16 @@
-'use strict'
-
-const test = require('tape')
-const proxiquire = require('proxyquire')
+import proxiquire = require('proxyquire')
+import test = require('tape')
+import getProcessEnv from '../lib/getProcessEnv'
 
 const MockHttp = mockHttpAgent('http')
-MockHttp.HttpsAgent = mockHttpAgent('https')
-const agent = proxiquire('./index.js', {
+MockHttp['HttpsAgent'] = mockHttpAgent('https')
+const agent = proxiquire('../lib/index.js', {
   'agentkeepalive': MockHttp,
   'https-proxy-agent': mockHttpAgent('https-proxy')
-})
+}).default
 
-function mockHttpAgent (type) {
-  return function Agent (opts) {
+function mockHttpAgent (type: string) {
+  return function Agent (opts: any) { // tslint:disable-line:no-any
     return Object.assign({}, opts, { __type: type })
   }
 }
@@ -19,10 +18,10 @@ function mockHttpAgent (type) {
 test('extracts process env variables', t => {
   process.env = { TEST_ENV: 'test', ANOTHER_ENV: 'no' }
 
-  t.deepEqual(agent.getProcessEnv('test_ENV'), 'test', 'extracts single env')
+  t.deepEqual(getProcessEnv('test_ENV'), 'test', 'extracts single env')
 
   t.deepEqual(
-    agent.getProcessEnv(['not_existing_env', 'test_ENV', 'another_env']),
+    getProcessEnv(['not_existing_env', 'test_ENV', 'another_env']),
     'test',
     'extracts env from array of env names'
   )
@@ -31,21 +30,21 @@ test('extracts process env variables', t => {
 
 const OPTS = {
   agent: null,
-  maxSockets: 5,
   ca: 'ca',
   cert: 'cert',
   key: 'key',
   localAddress: 'localAddress',
+  maxSockets: 5,
   strictSSL: 'strictSSL',
-  timeout: 5
+  timeout: 5,
 }
 
 test('all expected options passed down to HttpAgent', t => {
   t.deepEqual(agent('http://foo.com/bar', OPTS), {
     __type: 'http',
-    maxSockets: 5,
     localAddress: 'localAddress',
-    timeout: 6
+    maxSockets: 5,
+    timeout: 6,
   }, 'only expected options passed to HttpAgent')
   t.end()
 })
@@ -56,10 +55,10 @@ test('all expected options passed down to HttpsAgent', t => {
     ca: 'ca',
     cert: 'cert',
     key: 'key',
-    maxSockets: 5,
     localAddress: 'localAddress',
+    maxSockets: 5,
     rejectUnauthorized: 'strictSSL',
-    timeout: 6
+    timeout: 6,
   }, 'only expected options passed to HttpsAgent')
   t.end()
 })
@@ -70,18 +69,18 @@ test('all expected options passed down to proxy agent', t => {
   }, OPTS)
   t.deepEqual(agent('https://foo.com/bar', opts), {
     __type: 'https-proxy',
-    host: 'my.proxy',
-    port: '1234',
-    protocol: 'https:',
-    path: '/foo',
     auth: 'user:pass',
     ca: 'ca',
     cert: 'cert',
+    host: 'my.proxy',
     key: 'key',
-    maxSockets: 5,
     localAddress: 'localAddress',
+    maxSockets: 5,
+    path: '/foo',
+    port: '1234',
+    protocol: 'https:',
     rejectUnauthorized: 'strictSSL',
-    timeout: 6
+    timeout: 6,
   }, 'only expected options passed to https proxy')
   t.end()
 })
