+++ packages/npm-registry-agent/typings/index.d.ts
@@ -1,0 +1,9 @@
+declare module 'https-proxy-agent' {
+  const anything: any;
+  export = anything;
+}
+
+declare module 'socks-proxy-agent' {
+  const anything: any;
+  export = anything;
+}
